\name{makeBasis.slearner}
\alias{makeBasis.slearner}
\title{
  Grow a deep neural network basis.
}
\description{
Given initial predictors and output variable, the function grows a deep neural network basis as described in the reference.
}

\usage{
  makeBasis.slearner(x, y, widths, control=makeControl())
}


\arguments{
  \item{x}{ A (numeric) design matrix.
}
  \item{y}{
  The output variable. Should be coerceable to a numeric vector.
}
  \item{widths}{
  Numeric vector with the number of nodes in each layer of the neural network.
}
  \item{control}{
  A list of control parameters generated by \code{\link{makeControl}}.
}
}


\details{
 The algorith used to grow the network is described in the reference.
}

\value{
A list including:
  \item{basis} A matrix object with the augmented basis.
  \item{makeBasis} A function. Allowing the augmentation of new design matrices.
}

\references{
  Livni, Roi, Shai Shalev-Shwartz, and Ohad Shamir. “A Provably Efficient Algorithm for Training Deep Networks.” arXiv:1304.7045 
}

\author{
  Jonathan Rosenblatt<john.ros.work@gmail.com> based on Matlab code by Ohad Shamir.
}


\seealso{
  \code{\link{svm.slearner}}
}

\examples{
 load(file='Package/data/test_data.RData')
 widths<- c(10,10)
 
 train.ind<- rep(FALSE, nrow(x))
 train.ind[1:250]<- TRUE
 .test<- makeBasis.slearner(x=test.data$X[train.ind,], y=test.data$Y[train.ind], widths=widths)
}


\keyword{ neural }
\keyword{ regression }
