\name{makeControl}
\alias{makeControl}

\title{
  Generate control object for \code{\link{makeBasis.slearner}}.
  Currently deprecatred.
}
\usage{
  makeControl()
}

\arguments{}

\description{
Currently deprecated.
}

\value{}

\references{
}

\author{
Jonathan Rosenblatt
}

\note{
  Made as provisions for future options.
}



\seealso{
  \code{\link{makeBasis.slearner}}
}

\examples{
}
